fix(shelf): one quantity, three columns, and the rounding between them
`scripts/verify_shelf_domain.py` demonstrated it against the real dispense
path: 2.9 units off a placement of 30 should leave 27.1, and left

    shelf_placements.units          25   (-2.1)
    pharmacy_shelves.current_units  29   (+1.9)
    shelf_transfer_events.delta     -2   (the audit row, rounded too)

Two copies of one number four apart, and an audit trail that could not
reconstruct either. All three columns were INTEGER and `_take_off_shelf`
converted the allocator's exact Decimal a different way into each —
float(take.after), int(take.units), -int(take.units). `shelf.allocate` was
never at fault; the loss was entirely in the write.

It matters because the shelf ledger is the "expected" side of the theft
detector in shelf.py:reconcile. The placement copy under-reports the floor, so
a correct count reads as SURPLUS and real loss is never investigated; the
cached copy over-reports it, so the same count reads as MISSING and the system
accuses staff over its own arithmetic. Liquids, creams and paediatric doses are
ordinary, so this was the common path.

Migration 0054 widens all three to Numeric(10,3), matching
inventory_lots.quantity_on_hand and prescription_fills.quantity_dispensed, and
the write path passes the Decimal through. Neither half works alone.
capacity_units stays INTEGER: it describes the furniture, not a quantity
anything divides.

The GREATEST(0, ...) clamp on current_units is gone. It wrote a plausible zero
whenever the cache could not cover the take — the fallback-constant pattern the
provenance rule forbids, leaving the shelf reading empty with no row saying why.
A cache that cannot cover its own take has already drifted, so the real
arithmetic is stored (negative if that is the truth) and flagged on the returned
take and on the durable movement row. A negative current_units is diagnostic.

depot_transfer's increment path had the same int() truncation and would have
rounded a fractional cached total away on the very next placement.

Left open and referred, not decided: current_units is a second writable copy of
SUM(shelf_placements.units). Exact arithmetic stops it drifting by accident, not
by omission. docs/design/SHELF_LEDGER.md section 5 states the options.

Verified on a disposable clone; the shared pharmpilot_test was never migrated.
Migration proven up->down->up. verify_shelf_domain.py: PASS, 0 findings (was 3
notes). tests/unit: 2178 passed, 1 pre-existing order-dependent failure that
passes in isolation.

+++ shared/models/depot.py
@@ -28,7 +28,16 @@ class PharmacyShelf(AuditedBase):
     label: Mapped[str] = mapped_column(String(40), nullable=False)        # e.g. "A-03-B-05"
     zone: Mapped[str | None] = mapped_column(String(40), nullable=True)
     capacity_units: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    current_units: Mapped[int] = mapped_column(Integer, nullable=False, default=0)  # denormalized cache of Σ placements
+    # Numeric(10,3), matching inventory_lots.quantity_on_hand and
+    # prescription_fills.quantity_dispensed (migration 0054). As an INTEGER this
+    # rounded away every fractional dispense; capacity_units above stays whole
+    # because it describes the furniture, not a quantity anything divides.
+    #
+    # A denormalised cache of Σ placements, and therefore a SECOND WRITABLE COPY
+    # of a derivable number. Exact arithmetic stops it drifting by accident; it
+    # does not stop a write path forgetting to update it. Referred to the
+    # inventory owner — see docs/design/SHELF_LEDGER.md, "The cache".
+    current_units: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False, default=0)
     storage_condition: Mapped[str] = mapped_column(String(20), nullable=False, default="ROOM_TEMP")
     is_active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
 
@@ -40,7 +49,9 @@ class ShelfPlacement(AuditedBase):
     inventory_lot_id: Mapped[UUID] = mapped_column(ForeignKey("inventory_lots.id"), nullable=False, index=True)
     shelf_id: Mapped[UUID] = mapped_column(ForeignKey("pharmacy_shelves.id"), nullable=False, index=True)
     ndc11: Mapped[str] = mapped_column(String(11), nullable=False, index=True)
-    units: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    # Numeric(10,3) since migration 0054: what stands on a shelf is measured in
+    # the same units as the lot it came from, and half a bottle is ordinary.
+    units: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False, default=0)
     placed_by: Mapped[UUID | None] = mapped_column(PGUUID(as_uuid=True), nullable=True)
     placed_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
 
@@ -66,7 +77,10 @@ class ShelfTransferEvent(AuditedBase):
     inventory_lot_id: Mapped[UUID] = mapped_column(ForeignKey("inventory_lots.id"), nullable=False)
     shelf_id: Mapped[UUID] = mapped_column(ForeignKey("pharmacy_shelves.id"), nullable=False)
     ndc11: Mapped[str] = mapped_column(String(11), nullable=False)
-    quantity_delta: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Numeric(10,3) since migration 0054. This is the audit trail for shelf
+    # movement; as an INTEGER a fractional take was recorded as 0, so the one
+    # row that could have reconstructed the drift was rounded away itself.
+    quantity_delta: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False)
     performed_by: Mapped[UUID | None] = mapped_column(PGUUID(as_uuid=True), nullable=True)
     barcode_verification_result: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
     ai_verification_result: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
